Flytt @Transactional til egen komponent
Metoden som var annotert med @Tranaactional ble kallt fra
en mode i samme klasse, noe som gjør at annoteringen ikke slår til
og dermed at koden ikke kjørte i en transaksjon.
